requirements.txt
Django>=6.0
channels>=4.0
channels-redis>=4.0
celery>=5.3
django-celery-results>=2.4
redis>=4.5
psycopg2-binary>=2.9
djangorestframework>=3.14
gunicorn>=20.1
whitenoise>=6.0
